test: remove unnecessary test
