+++ package.json
@@ -26,12 +26,12 @@
     ],
     "scripts": {
         "start": "NODE_ENV=production node dist/server.js",
-        "start-dev": "nodemon src/server.ts",
+        "start:dev": "nodemon src/server.ts",
         "build": "rm -rf build && tsc",
         "lint": "eslint ./src --fix",
         "format": "prettier --write ./",
         "test": "NODE_ENV=test jest --config ./jest.config.ts",
-        "test-dev": "NODE_ENV=test jest --config ./jest.config.ts --watchAll "
+        "test:dev": "NODE_ENV=test jest --config ./jest.config.ts --watchAll "
     },
     "dependencies": {
         "@vonage/server-sdk": "^2.10.10",
